20200115 대댓글 구현

+++ Basic_Board_Project/src/com/service/comment/CommentServiceImpl.java
@@ -32,13 +32,26 @@ public class CommentServiceImpl implements CommentService {
 	}
 
 	@Override
+	public int insertComment(CommentsVO parentComment, CommentsVO subComment) {
+		return commentDAO.insertComment(parentComment, subComment);
+		
+	}
+
+	@Override
 	public int updateComment(CommentsVO comment) {
 		return commentDAO.updateComment(comment);
 	}
 
 	@Override
-	public int deleteComment(int COMMENT_CODE) {
-		return commentDAO.deleteComment(COMMENT_CODE);
+	public boolean deleteComment(int COMMENT_CODE, int GROUP_DEPTH, int GROUP_NO) {
+		
+		// 댓글(대댓글) 삭제 적합성 판단 (내 DEPTH와 내가 속한 GROUP의 MAX(DEPTH)가 같다면 삭제 가능)
+		if(commentDAO.maxDepth(GROUP_NO) == GROUP_DEPTH) {
+			commentDAO.deleteComment(COMMENT_CODE);
+			return true;
+		} else {
+			return false;
+		}
 	}
 
 	@Override
@@ -50,5 +63,4 @@ public class CommentServiceImpl implements CommentService {
 	public int badComment(int COMMENT_CODE) {
 		return commentDAO.badComment(COMMENT_CODE);
 	}
-
 }
